ArmPkg: Fix call to BuildCpuHob in CpuPei.c (#295)
## Description

Fix CpuPei.c for aarch64. Change function signature for
ArmGetPhysicalAddressBits.

- [ ] Impacts functionality?
- [ ] Impacts security?
- [ ] Breaking change?
- [ ] Includes tests?
- [ ] Includes documentation?

## How This Was Tested

Validated handoff on virtual platform bringup

## Integration Instructions

N/A

---------

+++ ArmPkg/Drivers/CpuPei/CpuPei.c
@@ -135,7 +135,7 @@ InitializeCpuPeim (
   ARM_CORE_INFO         *ArmCoreInfoTable;
 
   // Publish the CPU memory and io spaces sizes
-  BuildCpuHob (ArmGetPhysicalAddressBits (), PcdGet8 (PcdPrePiCpuIoSize));
+  BuildCpuHob ((UINT8)ArmGetPhysicalAddressBits (), PcdGet8 (PcdPrePiCpuIoSize)); // MU_CHANGE Add typecast
 
   // Only MP Core platform need to produce gArmMpCoreInfoPpiGuid
   Status = PeiServicesLocatePpi (&gArmMpCoreInfoPpiGuid, 0, NULL, (VOID **)&ArmMpCoreInfoPpi);
